+++ src/plugin-sdk/command-auth.ts
@@ -101,8 +101,8 @@ export {
   resolveModelsCommandReply,
 } from "../auto-reply/reply/commands-models.js";
 export type { ModelsProviderData } from "../auto-reply/reply/commands-models.js";
-export { resolveStoredModelOverride } from "../auto-reply/reply/stored-model-override.js";
-export type { StoredModelOverride } from "../auto-reply/reply/stored-model-override.js";
+export { resolveStoredModelOverride } from "../sessions/stored-model-overrides.js";
+export type { StoredModelOverride } from "../sessions/stored-model-overrides.js";
 
 /**
  * Inputs for legacy sender command authorization.
